Namespacing customized rendering classes
