+++ .rubocop.yml
@@ -38,7 +38,7 @@ Metrics/ClassLength:
 
 Rails/OutputSafety:
   Exclude:
-    - 'app/presentors/custom_join.rb'
+    - 'app/presenters/rendering/*'
 
 Rails/Output:
   Exclude:
