+++ config/initializers/blacklight.rb
@@ -1,5 +1,5 @@
 Blacklight::Rendering::Pipeline.operations = [Blacklight::Rendering::HelperMethod,
                                               Blacklight::Rendering::LinkToFacet,
                                               Blacklight::Rendering::Microdata,
-                                              PreserveNewLines,
-                                              CustomJoin]
+                                              Rendering::PreserveNewLines,
+                                              Rendering::CustomJoin]
